Re-organize project structure.
