+++ pkg/jwt/jwt_test/key_test.go
@@ -1,21 +1,20 @@
-package jwt_auth_test
+package jwt_test
 
 import (
-	"testing"
-
-	"eau-de-go/internal/jwt_auth"
+	"eau-de-go/pkg/jwt"
 	"github.com/stretchr/testify/assert"
+	"testing"
 )
 
 func TestGetInMemoryRsaKeyPair_Singleton(t *testing.T) {
-	keyPair1 := jwt_auth.GetInMemoryRsaKeyPair()
-	keyPair2 := jwt_auth.GetInMemoryRsaKeyPair()
+	keyPair1 := jwt.GetInMemoryRsaKeyPair()
+	keyPair2 := jwt.GetInMemoryRsaKeyPair()
 
 	assert.Equal(t, keyPair1, keyPair2, "GetInMemoryRsaKeyPair should always return the same instance")
 }
 
 func TestInMemoryRsaKeyPair_GetVerificationKey(t *testing.T) {
-	keyPair := jwt_auth.GetInMemoryRsaKeyPair()
+	keyPair := jwt.GetInMemoryRsaKeyPair()
 
 	verificationKey1, err1 := keyPair.GetVerificationKey()
 	assert.Nil(t, err1, "GetVerificationKey should not return an error")
@@ -27,7 +26,7 @@ func TestInMemoryRsaKeyPair_GetVerificationKey(t *testing.T) {
 }
 
 func TestInMemoryRsaKeyPair_GetSigningKey(t *testing.T) {
-	keyPair := jwt_auth.GetInMemoryRsaKeyPair()
+	keyPair := jwt.GetInMemoryRsaKeyPair()
 
 	signingKey1, err1 := keyPair.GetSigningKey()
 	assert.Nil(t, err1, "GetSigningKey should not return an error")
