+++ pkg/jwt/key.go
@@ -1,4 +1,4 @@
-package jwt_auth
+package jwt
 
 import (
 	"crypto/rand"
