+++ pkg/password/password.go
@@ -1,4 +1,4 @@
-package utils
+package password
 
 import (
 	"golang.org/x/crypto/bcrypt"
